DemandMaps use Doubles, refactored trip generation, DemandMapCollector
Converted most DemandMap implementations to use doubles rather than
floats
Created DemandMapCollector and CombinedDemandMap classes, now in use in
AggregatePAMatrixCollector
TripGeneration now uses DemandMaps where it should, instead of Maps from
TSZs to Doubles
Lots of cleanup in main script

+++ src/edu/utexas/wrap/assignment/bush/Bush.java
@@ -228,7 +228,7 @@ public class Bush implements AssignmentContainer {
 			if (bv == null) continue;
 			TravelSurveyZone tsz = bv.getHead().getZone();
 			if (tsz == null) continue;
-			Float x = demand.getOrDefault(tsz, 0.0F);
+			Double x = demand.getOrDefault(tsz, 0.0);
 			if (x <= 0.0) continue;
 			dumpFlow(tsz.getNode(),x.doubleValue());	//recursively push flow onto the bush
 		}
@@ -361,7 +361,7 @@ public class Bush implements AssignmentContainer {
 	 * @see edu.utexas.wrap.assignment.AssignmentContainer#getDemand(edu.utexas.wrap.net.Node)
 	 */
 	public Float getDemand(Node node) {
-		return demand.get(node.getZone());
+		return demand.get(node.getZone()).floatValue();
 	}
 	
 	/**
@@ -1302,7 +1302,7 @@ public class Bush implements AssignmentContainer {
 				//And which doesn't originate from this node
 				.filter(e -> !e.getKey().getTail().equals(origin.getNode()))
 				//And has more demand on it than the destination demand
-				.filter(e -> e.getValue() > demand.getOrDefault(e.getKey().getHead().getZone(), 0.0f))
+				.filter(e -> e.getValue() > demand.getOrDefault(e.getKey().getHead().getZone(), 0.0))
 				//Then something is wrong and this connector is allowing through too much demand
 				.findAny().isPresent()) throw new RuntimeException("Invalid Centroid Connector usage in Bush");
